[+] A change which allows for username input
the original, only hides "root". Of course, this can be changed from  "root" to "alice" (if the user is alice, for example to **hide the user alice**). But, I wanted to make it a bit more flexible, or a bit more easy to use when compiled. So I add a input feature, where you can at the time of the LKM Insert; specify and pass a username (like, for example "alice"). I got inspiration from [EXEC-LKM](https://github.com/loneicewolf/EXEC_LKM/blob/main/LKM/lkm_exec.c)